// frontend/src/shared/api/generated/model/pastPaperListOut.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * novalearn-api
 * OpenAPI spec version: 0.1.0
 */

export interface PastPaperListOut {
  id: string;
  user_id: string | null;
  folder_id: string;
  name: string;
  original_filename: string | null;
  mark_scheme_filename?: string | null;
  status: string;
  processing_phase?: string | null;
  created_at: string;
}